+++ app/webFrontend/src/app/auth/register/register.component.ts
@@ -7,6 +7,7 @@ import {MatSnackBar} from '@angular/material';
 import {errorCodes} from '../../../../../../api/src/config/errorCodes';
 import {TitleService} from '../../shared/services/title.service';
 import {APIInfoService} from '../../shared/services/data.service';
+import {emailValidator} from '../../shared/validators/validators';
 
 @Component({
   selector: 'app-register',
@@ -112,7 +113,7 @@ export class RegisterComponent implements OnInit {
         firstName: ['', Validators.compose([Validators.required, Validators.minLength(2), Validators.maxLength(64)])],
         lastName: ['', Validators.compose([Validators.required, Validators.minLength(2), Validators.maxLength(64)])],
       }),
-      email: ['', Validators.compose([Validators.required, Validators.email, this.validateTeacherEmail.bind(this)])],
+      email: ['', Validators.compose([emailValidator, Validators.required, Validators.email, this.validateTeacherEmail.bind(this)])],
       uid: ['', [this.validateMatriculationNumber.bind(this)]]
     });
   }